requirements.txt
nibabel~=3.2.1
pycimg~=1.0.2
tensorflow-gpu==1.15
tensorflow-probability==0.7.0
matplotlib==3.1.0
colorcet==2.0.1
brokenaxes==0.3.1
sacred==0.7.5
pymongo==3.8.0